aider: Center all the pitch control lines and change `css/read-a-sentence/style.css`.
# Aider chat conversation:

USER: center all the pitch controls lines (i.e. label, buttons and text input) (also change `css/read-a-sentence/style.css`

ASSISTANT: To center all the pitch control lines, we will add a CSS rule to the `.pitch-control` class in the `css/read-a-sentence/style.css` file to center the content, and then we will ensure that each `pitchControl` div in the `read-a-sentence.html` file has this class applied.

First, let's add the CSS rule to center the content within the `.pitch-control` class:

```css
css/read-a-sentence/style.css
<<<<<<< SEARCH

  /* Container for the elements */
  .container {
    display: flex;
    flex-direction: column;
  }
=======

  .pitch-control {
    text-align: center;
  }

  /* Container for the elements */
  .container {
    display: flex;
    flex-direction: column;
  }
>>>>>>> REPLACE
```

Next, we will ensure that the `pitchControl` divs have the `pitch-control` class applied. Since the `read-a-sentence.html` file already has a `pitch-controls` ID on the container for these controls, we can use that to apply the styles. However, if you want to apply the class directly to each `pitchControl` div, here's how you would do it:

```html
read-a-sentence.html
<<<<<<< SEARCH
    var pitchControl = document.createElement('div');
=======
    var pitchControl = document.createElement('div');
    pitchControl.className = 'pitch-control';
>>>>>>> REPLACE
```

This change will apply the `pitch-control` class to each `pitchControl` div, which will center the label, buttons, and text input as per the new CSS rule we added. If the `pitch-controls` ID is meant to be used for styling, you can alternatively add the text-align property to its CSS rule instead of creating a new class.